requirements.txt
pandas>=2.2
numpy>=2.0
scikit-learn>=1.5
matplotlib>=3.8
seaborn>=0.13
joblib>=1.4
PyYAML>=6.0
ydata-profiling>=4.12
shap>=0.46
featuretools>=1.31
pytest>=8.0
setuptools>=65.0
